requirements.txt
fastapi
uvicorn[standard]
pinecone
groq
python-dotenv
pydantic
cohere
